Fixes to compile on linux

+++ engine/FullscreenRendering_OpenGL.cpp
@@ -88,8 +88,8 @@ namespace jle
 		glDisable(GL_DEPTH_TEST);
 
 		// Get the texture from the framebuffer
-		glBindTexture(GL_TEXTURE_2D, (unsigned int)framebuffer.GetTexture());
-		glStaticState.globalActiveTexture = (unsigned int)framebuffer.GetTexture();
+        glBindTexture(GL_TEXTURE_2D,framebuffer.GetTexture());
+        glStaticState.globalActiveTexture = framebuffer.GetTexture();
 
 		// Draw quad with framebuffer's texture over the entire screen
 		glDrawArrays(GL_TRIANGLES, 0, 6);
